fix & optimize: harden core PCC logic, vectorize graph construction, and remove degree caps
- **Fixes**:
  - Fixed a critical NameError in `pcc_numba.py` (call to non-existent `_pcc_step_numba`).
  - Corrected integer type inconsistency in `pcc_numpy.py` (`int32` -> `int64` for neighbor indexing).
  - Cleaned up unused variables and redundant array order flags in `pcc.py`.
  - Removed duplicate imports in `pcc_numba.py`.

- **Performance**:
  - Fully vectorized `neib_list` construction in `pcc_graph.py`, eliminating O(N) Python loops to ensure ordens of magnitude faster graph building on large datasets.
  - Refactored the innermost Cython loop in `pcc_step.pyx` to use direct memoryview indexing.

- **Hardening / Cap Removal**:
  - Removed the artificial `max_degree=500` limit in `build_graph_from_edge_index` to support connectivity in dense graphs.
  - Eliminated the legacy `max_buf_size` artificial cap in the Cython kernel.

+++ pcc_numpy.py
@@ -89,7 +89,7 @@ def pcc_step_numpy(neib_list, neib_qt,
         no_prob = ~has_prob
         if np.any(no_prob):
             np_idx = np.where(no_prob)[0]
-            rand_choices = (np.random.random(len(np_idx)) * k_vals_g[np_idx]).astype(np.int32)
+            rand_choices = (np.random.random(len(np_idx)) * k_vals_g[np_idx]).astype(np.int64)
             next_nodes[g_indices[np_idx]] = neighbors_g[np_idx, rand_choices]
 
     # 4. Process Random Walks
@@ -99,7 +99,7 @@ def pcc_step_numpy(neib_list, neib_qt,
         k_vals_r = neib_qt[cur_nodes_r]
         
         # Pick random neighbor index
-        rand_choices = (np.random.random(len(r_indices)) * k_vals_r).astype(np.int32)
+        rand_choices = (np.random.random(len(r_indices)) * k_vals_r).astype(np.int64)
         # Efficiently extract selected neighbors
         # We can't easily 2D index if rows have different k, but here next_nodes 
         # is just neib_list[cur_node, choice]
